requirements.txt
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
branca==0.7.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
duckdb==1.0.0
executing==2.0.1
folium==0.17.0
fonttools==4.53.1
h11==0.14.0
htmltools==0.5.3
idna==3.7
ipyleaflet==0.19.2
ipython==8.18.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
jupyter-leaflet==0.19.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt-toolkit==3.0.36
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
pywin32==306
questionary==2.0.1
requests==2.32.3
setuptools==72.1.0
shiny==1.0.0
shinywidgets==0.3.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.2
traitlets==5.14.3
traittypes==0.2.1
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
uvicorn==0.30.5
watchfiles==0.23.0
wcwidth==0.2.13
websockets==12.0
widgetsnbextension==4.0.11
xyzservices==2024.6.0
